requirements.txt
fastapi==0.68.2
greenlet==1.1.2
pathspec==0.9.0
pydantic==1.9.0
python-dotenv==0.20.0
starlette==0.14.2
typing-extensions==4.2.0
uvicorn==0.30.1
PyJWT==2.3.0
cryptography==36.0.1
requests==2.21.0
python-dateutil
Unidecode
python-slugify
pymongo
pydantic[email]
pymoises
ruff
colorama
pyfiglet